prog(): Added intermediate board setup in initialize.js
When creating a board users are sent to the 'initialize' route. Users input their desired tile labels and on submit they are routed to the /play/[slug] route where the board will no longer be editable. Routing is currently working as expected and rooms stay active as the browser redirects.

TODO: Change Board component so it is no longer editable. Add Ably messages to [slug].js to update individual users board states.